added support for braceexpansion

+++ utils/general.py
@@ -518,7 +518,6 @@ def check_font(font=FONT, progress=False):
 
 def check_dataset(data, autodownload=True):
     """Validates and/or auto-downloads a dataset, returning its configuration as a dictionary."""
-
     # Download (optional)
     extract_dir = ""
     if isinstance(data, (str, Path)) and (is_zipfile(data) or is_tarfile(data)):
@@ -557,27 +556,28 @@ def check_dataset(data, autodownload=True):
     train, val, test, s = (data.get(x) for x in ("train", "val", "test", "download"))
     if val:
         val = [Path(x).resolve() for x in (val if isinstance(val, list) else [val])]  # val path
-        if not all(x.exists() for x in val):
-            LOGGER.info("\nDataset not found ⚠️, missing paths %s" % [str(x) for x in val if not x.exists()])
-            if not s or not autodownload:
-                raise Exception("Dataset not found ❌")
-            t = time.time()
-            if s.startswith("http") and s.endswith(".zip"):  # URL
-                f = Path(s).name  # filename
-                LOGGER.info(f"Downloading {s} to {f}...")
-                torch.hub.download_url_to_file(s, f)
-                Path(DATASETS_DIR).mkdir(parents=True, exist_ok=True)  # create root
-                unzip_file(f, path=DATASETS_DIR)  # unzip
-                Path(f).unlink()  # remove zip
-                r = None  # success
-            elif s.startswith("bash "):  # bash script
-                LOGGER.info(f"Running {s} ...")
-                r = subprocess.run(s, shell=True)
-            else:  # python script
-                r = exec(s, {"yaml": data})  # return None
-            dt = f"({round(time.time() - t, 1)}s)"
-            s = f"success ✅ {dt}, saved to {colorstr('bold', DATASETS_DIR)}" if r in (0, None) else f"failure {dt} ❌"
-            LOGGER.info(f"Dataset download {s}")
+        if 'webdataset' not in data.keys():
+            if not all(x.exists() for x in val):
+                LOGGER.info("\nDataset not found ⚠️, missing paths %s" % [str(x) for x in val if not x.exists()])
+                if not s or not autodownload:
+                    raise Exception("Dataset not found ❌")
+                t = time.time()
+                if s.startswith("http") and s.endswith(".zip"):  # URL
+                    f = Path(s).name  # filename
+                    LOGGER.info(f"Downloading {s} to {f}...")
+                    torch.hub.download_url_to_file(s, f)
+                    Path(DATASETS_DIR).mkdir(parents=True, exist_ok=True)  # create root
+                    unzip_file(f, path=DATASETS_DIR)  # unzip
+                    Path(f).unlink()  # remove zip
+                    r = None  # success
+                elif s.startswith("bash "):  # bash script
+                    LOGGER.info(f"Running {s} ...")
+                    r = subprocess.run(s, shell=True)
+                else:  # python script
+                    r = exec(s, {"yaml": data})  # return None
+                dt = f"({round(time.time() - t, 1)}s)"
+                s = f"success ✅ {dt}, saved to {colorstr('bold', DATASETS_DIR)}" if r in (0, None) else f"failure {dt} ❌"
+                LOGGER.info(f"Dataset download {s}")
     check_font("Arial.ttf" if is_ascii(data["names"]) else "Arial.Unicode.ttf", progress=True)  # download fonts
     return data  # dictionary
